# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "auctionEngine API",
    "version": "1.0.0",
    "description": "Live auction and bidding backend. Amounts are integers in the smallest currency unit. Errors are {\"error\": \"...\"}; validation errors add a fields map. Most endpoints are rate limited (429 with Retry-After)."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "invalid input"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "item.name": "is required"
            }
          }
        }
      },
...
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "CreateAuction": {
        "type": "object",
        "required": [
          "item",
          "starting_price",
          "starts_at",
          "ends_at"
        ],
        "properties": {
          "item": {
            "type": "object",
            "required": [
              "name",
              "type"
            ],
            "properties": {
              "name": {
                "type": "string",
                "maxLength": 200
              },
              "type": {
                "type": "string",
                "maxLength": 50
              },
              "description": {
                "type": "string",
                "maxLength": 5000
              }
            }
          },
          "starting_price": {
            "type": "integer",
            "minimum": 0
          },
          "min_increment": {
            "type": "integer",
            "minimum": 0,
            "description": "0 means 1"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Auction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid"
          },
          "item": {
            "$ref": "#/components/schemas/Item"
          },
          "starting_price": {
            "type": "integer"
          },
          "min_increment": {
            "type": "integer"
          },
          "current_bid": {
            "type": [
              "integer",
              "null"
            ]
          },
          "current_bidder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "bid_count": {
            "type": "integer"
          },
          "extensions": {
            "type": "integer"
          },
          "settled_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_ACTIVE",
              "ACTIVE",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/auctions": {
      "post": {
        "tags": [
          "auctions"
        ],
        "summary": "List an item for auction",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuction"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Auction"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "auctions"
        ],
        "summary": "List auctions, newest first (keyset pagination)",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "NOT_ACTIVE",
                "ACTIVE",
                "COMPLETED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "owner",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "me"
              ]
            },
            "description": "Requires a token"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Opaque; pass next_cursor from the previous page."
          }
        ],
        "responses": {
          "200": {
            "description": "A page",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "auctions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Auction"
                      }
                    },
                    "next_cursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "owner=me without a token, or an invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}